+++ db/migrate/20141117142904_add_verifying_sheet_id_to_sheets.rb
@@ -1,0 +1,5 @@
+class AddVerifyingSheetIdToSheets < ActiveRecord::Migration
+  def change
+    add_column :sheets, :verifying_sheet_id, :integer
+  end
+end